New module in core and made things better.

+++ setup/source/app.d
@@ -1,7 +1,7 @@
 #!/bin/env rdmd
 
 /// A helper script that automates the project setup.
-/// This script is designed with the idea that you use DUB.
+/// This script is designed with the idea that you use DUB, but it can work without DUB too.
 
 import std.format;
 import std.path;
@@ -132,10 +132,7 @@ void deleteFile(const(char)[] path) {
 }
 
 int main(string[] args) {
-    if (check(dubFile, false)) {
-        writeln("Error: This is not a DUB project.");
-        return 1;
-    }
+    auto isDUBProject = !check(dubFile, false);
     // Skip if the folders already exist.
     if (!check(assetsDir, false) || !check(webDir, false)) {
         writeln("Skipping setup because some folders already exist.");
@@ -145,28 +142,34 @@ int main(string[] args) {
     // Use the raylib-d script to download the raylib library files.
     // We also have to use `spawnShell` here because raylib-d:install does not accept arguments.
     // TODO: Ask the raylib-d project to do something about that.
-    auto canDownload = args.length == 1 || (args.length > 1 && args[1] != noLibsArg);
-    if (canDownload) {
-        run("dub add raylib-d");
-        writeln();
-        writeln(`"Saying yes to happiness means learning to say no to the things and people that stress you out." - Thema Davis`);
-        writeln();
-        auto pid = spawnShell("dub run raylib-d:install");
-        wait(pid);
-    } else if (args.length > 1 && args[1] != noLibsArg) {
-        writeln("Info: Pass `%s` if you don't want to download raylib.".format(noLibsArg));
+    if (isDUBProject) {
+        auto canDownload = args.length == 1 || (args.length > 1 && args[1] != noLibsArg);
+        if (canDownload) {
+            run("dub add raylib-d");
+            writeln();
+            writeln(`"Saying yes to happiness means learning to say no to the things and people that stress you out." - Thema Davis`);
+            writeln();
+            auto pid = spawnShell("dub run raylib-d:install");
+            wait(pid);
+        } else if (args.length > 1 && args[1] != noLibsArg) {
+            writeln("Info: Pass `%s` if you don't want to download raylib.".format(noLibsArg));
+        }
     }
 
     // Delete the old files.
-    deleteFile(dubFile);
-    deleteFile(dubyFile);
+    if (isDUBProject) {
+        deleteFile(dubFile);
+        deleteFile(dubyFile);
+        deleteFile(appFile);
+    }
     deleteFile(gitignoreFile);
-    deleteFile(appFile);
 
     // Create the new files.
-    std.file.write(dubFile, defaultDUBContent);
+    if (isDUBProject) {
+        std.file.write(dubFile, defaultDUBContent);
+        std.file.write(appFile, defaultAppContent);
+    }
     std.file.write(gitignoreFile, defaultGitignoreContent);
-    std.file.write(appFile, defaultAppContent);
     std.file.mkdir(assetsDir);
     std.file.mkdir(webDir);
     return 0;
